\name{bhpm.cluster.data1}
\docType{data}
\alias{bhpm.cluster.data1}
\title{Cluster analysis data.}
\description{
  This dataset contains counts of the outcomes and exposure times for two treatment groups and six clusters.
}
\usage{data(bhpm.cluster.data1)}
\format{A dataframe with columns \emph{Cluster}, \emph{Outcome.Grp}, \emph{Outcoome}, \emph{Trt.Grp}, \emph{Count}, \emph{Exposure}.
The dataframe contains 1860 observations.}
\keyword{datasets}
