\name{bhpm.cluster.data2}
\docType{data}
\alias{bhpm.cluster.data2}
\title{Cluster analysis data.}
\description{
  This dataset contains counts of the coutcome and exposures for two treatments and ten clusters.
}
\usage{data(bhpm.cluster.data2)}
\format{A dataframe with columns \emph{Cluster}, \emph{Outcome.Grp}, \emph{Outcome}, \emph{Trt.Grp}, \emph{Count}, \emph{Exposure}.
The dataframe contains 1860 observations.}
\keyword{datasets}
