\name{bhpm.multi.treatments}
\docType{data}
\alias{bhpm.multi.treatments}
\title{Cluster analysis data.}
\description{
  This data set contains counts of outcomes and exposure times for four treatments over all six clusters of an observational study.
}
\usage{data(bhpm.multi.treatments)}
\format{A dataframe with columns \emph{Cluster}, \emph{Outcome.Grp}, \emph{Outcome}, \emph{Trt.Grp}, \emph{Count}, \emph{Exposure}.
The dataframe contains 3720 observations.}
%\source{World Almanac and Book of Facts, 1975, page 406.}
%\references{
%  McNeil, D. R. (1977) \emph{Interactive Data Analysis}.
%  New York: Wiley.
%}
\keyword{datasets}
