\name{bhpm-package}
\alias{bhpm-package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Bayesian Hierarchical Possion Models for Mulitple Grouped Outcomes with Clustering}
\description{
This package implements Bayesian Hierarchical models for the detection of differences in rates of outcomes for multiple treatments for clustered observations using groupings of outcomes.
}

\details{
The methods implemented use assumed groupings of outcomes to 
detect differences in the occurrence of the outcome for different treatments. Methods based on Bayesian Hierarchical
models are provided.
The methods are extension of those described in Carragher (2017) for multiple treatments and clusters.
}
\author{
R. Carragher<raymond.carragher@strath.ac.uk; rcarragh@gmail.com>
}

\references{
Carragher R. Detection of Safety Signals in Randomised Controlled Trials. PhD thesis. University of Strathclyde, 2017 <https://ethos.bl.uk/OrderDetails.do?uin=uk.bl.ethos.736866>.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Bayesian Hierarchy}
\keyword{Body-system}
\keyword{System organ class}
\keyword{bhpm-package}
\keyword{Cluster} % __ONLY ONE__ keyword per line
\keyword{Adverse event} % __ONLY ONE__ keyword per line
\keyword{Adverse outcome} % __ONLY ONE__ keyword per line
