\name{demo.cluster.data}
\docType{data}
\alias{demo.cluster.data}
\title{Cluster analysis data.}
\description{
  This dataset contains counts of the outcomes and exposure times for four treatment groups and three clusters.
}
\usage{data(demo.cluster.data)}
\format{A dataframe with columns \emph{Cluster},  \emph{Trt.Grp}, \emph{Outcome.Grp}, \emph{Outcome}, \emph{Exposure}, \emph{Count}.
The dataframe contains 108 observations.}
\keyword{datasets}
