\name{bhpm.multi.treatments.random.order}
\docType{data}
\alias{bhpm.multi.treatments.random.order}
\title{Cluster analysis data.}
\description{
  This data set contains count of outcome for four treatments over six clusters of an observational study. The data are in random order.
}
\usage{data(bhpm.multi.treatments.random.order)}
\format{A dataframe with columns \emph{Cluster}, \emph{Outcome.Grp}, \emph{Outcome}, \emph{Trt.Grp}, \emph{Count}, \emph{Exposure}.
The dataframe contains 3720 observations.}
\keyword{datasets}
